moving _getSubmittedTransaction off of web3Service into its v0 implementation
