+++ unlock-js/src/__tests__/web3Service.test.js
@@ -152,73 +152,6 @@ describe('Web3Service', () => {
       web3Service._getPastTransactionsForContract(contract, events, filter)
     })
   })
-
-  describe('_getSubmittedTransaction', () => {
-    const blockNumber = 29
-    const defaults = null
-
-    beforeEach(() => {
-      web3Service._watchTransaction = jest.fn()
-    })
-
-    it('should watch the transaction', done => {
-      expect.assertions(1)
-      web3Service.on('transaction.updated', () => {
-        expect(web3Service._watchTransaction).toHaveBeenCalledWith(
-          transaction.hash
-        )
-        done()
-      })
-
-      web3Service._getSubmittedTransaction(
-        transaction.hash,
-        blockNumber,
-        defaults
-      )
-    })
-
-    it('should emit a transaction.updated event with the right values', done => {
-      expect.assertions(4)
-      web3Service.on('transaction.updated', (hash, update) => {
-        expect(hash).toBe(transaction.hash)
-        expect(update.status).toEqual('submitted')
-        expect(update.confirmations).toEqual(0)
-        expect(update.blockNumber).toEqual(Number.MAX_SAFE_INTEGER)
-        done()
-      })
-      web3Service._getSubmittedTransaction(
-        transaction.hash,
-        blockNumber,
-        defaults
-      )
-    })
-
-    it('should invoke parseTransactionFromInput if the defaults include an input value', done => {
-      expect.assertions(3)
-
-      const defaults = {
-        input:
-          '0x2bc888bf00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000278d00000000000000000000000000000000000000000000000000002386f26fc10000000000000000000000000000000000000000000000000000000000000000000a',
-        to: '0xcfeb869f69431e42cdb54a4f4f105c19c080a601',
-      }
-
-      web3Service.parseTransactionFromInput = jest.fn(
-        (transactionHash, contract, transactionInput, address) => {
-          expect(transactionHash).toEqual(transaction.hash)
-          expect(transactionInput).toEqual(defaults.input)
-          expect(address).toEqual('0xcfeb869f69431e42cdb54a4f4f105c19c080a601')
-          done()
-        }
-      )
-
-      web3Service._getSubmittedTransaction(
-        transaction.hash,
-        blockNumber,
-        defaults
-      )
-    })
-  })
-
   describe('_getPendingTransaction', () => {
     const input =
       '0x2bc888bf00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000278d00000000000000000000000000000000000000000000000000002386f26fc10000000000000000000000000000000000000000000000000000000000000000000a'
